Don't enhance failure when running `git reset`
In this case, I don't think the extra indirection improves things. As a
matter of fact, it can make the error more confusing because of the
"maybe you misspelled it?" part, since in this case, the revision will
most likely be recorded in the lockfile but overwritten remotely, so the
user has not really typed it directly.